Split up inbox feature logic into separate files
